+++ tests/testthat/test-autoplot.R
@@ -13,6 +13,7 @@ result_classify <- hotspot_classify(
   cell_size = 0.01,
   quiet = TRUE
 )
+result_change <- hotspot_change(data_sf, quiet = TRUE)
 
 
 
@@ -28,6 +29,8 @@ test_that("error if `object` is not an SF object", {
   expect_error(autolayer(sf::st_drop_geometry(result_kde)))
   expect_error(autoplot(sf::st_drop_geometry(result_classify)))
   expect_error(autolayer(sf::st_drop_geometry(result_classify)))
+  expect_error(autoplot(sf::st_drop_geometry(result_change)))
+  expect_error(autolayer(sf::st_drop_geometry(result_change)))
 })
 
 test_that("error if `object` does not contain the required columns", {
@@ -37,6 +40,8 @@ test_that("error if `object` does not contain the required columns", {
   expect_error(autolayer(result_kde[, "geometry"]))
   expect_error(autoplot(result_classify[, "geometry"]))
   expect_error(autolayer(result_classify[, "geometry"]))
+  expect_error(autoplot(result_change[, "geometry"]))
+  expect_error(autolayer(result_change[, "geometry"]))
 })
 
 test_that("error if required column does not have correct type", {
@@ -46,6 +51,9 @@ test_that("error if required column does not have correct type", {
   result_kde$kde <- as.character(result_kde$kde)
   expect_error(autoplot(result_kde))
   expect_error(autolayer(result_kde))
+  result_change$change <- as.character(result_change$change)
+  expect_error(autoplot(result_change))
+  expect_error(autolayer(result_change))
 })
 
 
@@ -56,4 +64,5 @@ test_that("output has correct class", {
   expect_s3_class(autoplot(result_count), "ggplot")
   expect_s3_class(autoplot(result_kde), "ggplot")
   expect_s3_class(autoplot(result_classify), "ggplot")
+  expect_s3_class(autoplot(result_change), "ggplot")
 })